Updating trustee code: trustee_id from Long to int

+++ Backend/property-management/src/main/kotlin/com/example/propertymanagement/model/Trustee.kt
@@ -15,7 +15,7 @@ data class Trustee(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "trustee_id")
-    val trustee_id: Long = 0,
+    val trustee_id: Int = 0,
     val name: String,
     val email: String,
     val phone: String,
